% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{get_data}
\alias{get_data}
\title{Retrieve data from the ECB Statistical Data Warehouse API}
\usage{
get_data(key, filter = NULL)
}
\arguments{
\item{key}{A character string identifying the series to be retrieved}

\item{filter}{A named list with additional filters (see \code{details})}
}
\value{
A data frame
}
\description{
Retrieve data from the ECB Statistical Data Warehouse API
}
\examples{
# Get monthly data on annualized euro area headline HICP
hicp <- get_data("ICP.M.U2.N.000000.4.ANR")
head(hicp)
}

