% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_dataflows}
\alias{get_dataflows}
\title{Retrieve data frame of all datasets in the ECB Statistical Data Warehouse}
\usage{
get_dataflows(...)
}
\value{
A dataframe
}
\description{
#' @param ... Arguments passed to GET (e.g. timeout(10) to add maximum request time)
}
\examples{
df <- get_dataflows()
head(df)
}
