% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_description}
\alias{get_description}
\title{Get full, human-readable description of a series}
\usage{
get_description(key)
}
\arguments{
\item{key}{A character string identifying the series to be retrieved}
}
\value{
A character vector
}
\description{
Get full, human-readable description of a series
}
\examples{
get_description("ICP.M.DE.N.000000+XEF000.4.ANR")
}
